More cleanup in live driver tear-down (in the built-in drivers)
Add consistent 'dev' and 'id' options to each driver that allows 
specification of a specific hardware output device.  This allows 
addressing device by number, as well as makes sure the named-device 
option is available as 'dev' in each driver (PA used 'sink' and OSS used 
'dsp'). The old names are kept as synonyms.



git-svn-id: http://svn.xiph.org/trunk/ao@17629 0101bb08-14d6-0310-b084-bc0e0c8e3800