OS X driver patch from Kyle McKay:
use device names (or partial names) or, if you have it, the audio
device UID (UIDs look like
"AppleHDAEngineOutputDP:8,5,1,0:0:{2D4C-05ED-00000000}") with the
-d/--device option.  With the partial matching you can use "-d
speaker" or "-d headphones" or "-d hdmi" or "-d spdif" or "-d iMic"
etc. to select the output device.  (It's a case insensitive match.)

pull out the attempt to set the AudioHardware's CFRunLoop to NULL
because that's just wrong (and fortunately was being ignored by
CoreAudio)



git-svn-id: http://svn.xiph.org/trunk/ao@18881 0101bb08-14d6-0310-b084-bc0e0c8e3800