Replace autogen.sh with a simple wrapper about autoreconf.
The autoreconf tool does the same tool detection the older
shell script does, and is now sufficiently widely deployed
to depend on. The new script is what is currently used in
other Xiph.Org projects.

Note this is a change to the new pattern of needing to
invoke ./configure separately after running ./autogen.sh.