Revert "Insure that the alsa plugin links against libao, since the plugin calls ao_is_big_endian from libao."
This reverts commit d11c3f1acdecb8a400e8007e1ab17348b46fead3.

This change was the wrong way to 'fix' the problem reported in trac #1779.
In normal use the plugin should already have access to the symbols from
the library that loaded it, so creating a circular dependency between
them isn't helpful.

If some other user is dynamically loading the core libao itself as a
plugin to their application, then they need to use RTLD_GLOBAL to ensure
its symbols are made available to its plugins (in the same way that libao
needs to do this when dynamically loading alsa for alsa's own plugins).