pulse: fix possible overflow for buffer_time option
The int cast is wrong and unnecessary as internal->buffer_time is a
pa_usec_t (which is uint64_t). It does not even make sense to cast
it to a uint as it being multiplied by the sample rate (and then
divided by 1000000 before it get fitted into battr.tlength, which
is a uint32_t).

Also ditch the "+7" trick for format->bits as the function would
have returned already if it isn't 8/16(/24).